requirements.txt
asgiref==3.4.1
astroid==2.6.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.6
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
idna==3.2
isort==5.9.3
jwcrypto==1.0
kavenegar==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==1.7.1
pylint==2.9.6
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
